# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "湖南工学院创客系统API介绍",
        "title": "湖南工学院创客系统API",
        "termsOfService": "http://www.hnit.edu.cn",
        "contact": {
            "name": "adeljck",
            "url": "http://adeljck.github.io",
            "email": "fuck@you.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/api/v1",
    "paths": {
        "/academy": {
            "get": {
                "description": "学院列表",
                "produces": [
                    "application/json"
                ],
                "summary": "学院列表",
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":{\"item\":[],total:int},\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/enterpriseapply": {
            "post": {
                "description": "企业入驻",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "企业入驻",
                "parameters": [
                    {
                        "description": "enterprise",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.Enterprise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200, \"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"Tag Error\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/forget": {
            "post": {
                "description": "忘记密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "忘记密码",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserForget"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":{\"_id\":000000,\"token\":token,\"name\":adeljck},\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"Tag Error\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4006": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"unauthrized account\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/forget/changepassword": {
            "post": {
                "description": "忘记密码-\u003e修改密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "忘记密码-\u003e修改密码",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserForgetChangePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"Tag Error\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4006": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"unauthrized account\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/industry": {
            "get": {
                "description": "行业列表",
                "produces": [
                    "application/json"
                ],
                "summary": "行业列表",
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":{\"item\":[],total:int},\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "用户登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登陆",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserLoginService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":{\"_id\":000000,\"token\":token,\"name\":adeljck},\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"Tag Error\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4005": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"账号或密码错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4006": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"unauthrized account\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/mine/changepassword": {
            "post": {
                "description": "用户修改密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户修改密码",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserChangePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"Tag Error\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4006": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"unauthrized account\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "服务器状态检查",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "服务器状态检查",
                "responses": {
                    "200": {
                        "description": "{\"status\": 0,\"data\": null,\"msg\": \"Pong\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/project": {
            "put": {
                "description": "接手项目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "接手项目",
                "parameters": [
                    {
                        "description": "project",
                        "name": "projects",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.GetProject"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":bull,\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"failed\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "添加项目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "添加项目",
                "parameters": [
                    {
                        "description": "project",
                        "name": "projects",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ProjectAdd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":bull,\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"Tag Error\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除项目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除项目",
                "parameters": [
                    {
                        "description": "project",
                        "name": "projects",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ProjectDelete"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":bull,\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"failed\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/project/{id}": {
            "get": {
                "description": "获取项目详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取项目详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":items,\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"failed\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/projects/{page}": {
            "get": {
                "description": "项目列表",
                "produces": [
                    "application/json"
                ],
                "summary": "项目列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":{\"item\":[],total:int},\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/registe": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200,\"data\":null,\"msg\": \"success\" }",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4001": {
                        "description": "code\":4001,\"data\":null,\"msg\":\"JSON类型不匹配\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4002": {
                        "description": "code\":4002,\"data\":null,\"msg\":\"参数错误\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "4003": {
                        "description": "code\":4003,\"data\":null,\"msg\":\"Tag Error\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "serializer.PureErrorResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "object"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "serializer.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "object"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "service.Enterprise": {
            "type": "object",
            "required": [
                "companycode",
                "date",
                "email",
                "industry",
                "info",
                "legal",
                "legalid",
                "name",
                "phone",
                "site"
            ],
            "properties": {
                "companycode": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "legal": {
                    "type": "string"
                },
                "legalid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "site": {
                    "type": "string"
                }
            }
        },
        "service.GetProject": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "service.ProjectAdd": {
            "type": "object",
            "required": [
                "attach",
                "contact",
                "creator",
                "info",
                "money",
                "project_name",
                "short_info",
                "start_time",
                "status",
                "stop_time"
            ],
            "properties": {
                "attach": {
                    "type": "string"
                },
                "contact": {
                    "type": "object",
                    "$ref": "#/definitions/service.contact"
                },
                "creator": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "money": {
                    "type": "number"
                },
                "project_name": {
                    "type": "string"
                },
                "short_info": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "stop_time": {
                    "type": "string"
                }
            }
        },
        "service.ProjectDelete": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "service.User": {
            "type": "object",
            "required": [
                "academy",
                "class",
                "date",
                "email",
                "major",
                "name",
                "password",
                "phone",
                "repassword",
                "sid"
            ],
            "properties": {
                "academy": {
                    "type": "string"
                },
                "class": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "repassword": {
                    "type": "string"
                },
                "sid": {
                    "type": "string"
                }
            }
        },
        "service.UserChangePassword": {
            "type": "object",
            "required": [
                "oldpassword",
                "password",
                "repassword"
            ],
            "properties": {
                "oldpassword": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "repassword": {
                    "type": "string"
                }
            }
        },
        "service.UserForget": {
            "type": "object",
            "required": [
                "email",
                "phone",
                "sid"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "sid": {
                    "type": "string"
                }
            }
        },
        "service.UserForgetChangePassword": {
            "type": "object",
            "required": [
                "password",
                "repassword"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "repassword": {
                    "type": "string"
                }
            }
        },
        "service.UserLoginService": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "service.contact": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "qq": {
                    "type": "string"
                },
                "wechat": {
                    "type": "string"
                }
            }
        }
    }
}